Fix jelly-text support in AOT builds
Closes #128

Turns out that the Google proto classes require reflection for text serialization -- this was something that scalapb did not require.

To fix this, we could hardcode the methods and classes that need to have reflective access enabled, but it would be a mess to maintain. Instead, we have a custom Graal feature that discovers the needed classes and registers them for reflective access.