README: add note on JIT vs AOT perf (#157)
* README: add note on JIT vs AOT perf

True story. Observed on osm2rdf data, when converting ~7M triples from Turtle to Jelly. With JIT it took ~7s, while with AOT it was ~25–30s, depending on how the stars were aligning. Tested on the same machine as listed in Jelly benchmarks on the website.

The JIT throughput would be in the ballpark of 1MT/s, which is what we would expect from the Turtle parser (it tops out around this point). The AOT throughput is just abysmal.

* fix

* fix2