Initial implementation with context parameters (#162)
Closes #74

Everything came together really nicely with context parameters.

One thing is that I'm not sure how accurate the statistics are. I use getSerializedSize() of ProtoMessage as the source (and length of encoded string + 1 for the tag byte), but I'm not sure if these are correct.

I was hoping that triples_size would equal subject_size + predicate_size + object_size, but that isn't the case.